+++ packages/@aws-cdk/aws-eks-v2-alpha/lib/kubectl-handler/helm/__init__.py
@@ -52,6 +52,7 @@ def helm_handler(event, context):
     repository       = props.get('Repository', None)
     values_text      = props.get('Values', None)
     skip_crds        = props.get('SkipCrds', False)
+    take_ownership   = props.get('TakeOwnership', False)
 
     # "log in" to the cluster
     subprocess.check_call([ 'aws', 'eks', 'update-kubeconfig',
@@ -89,7 +90,7 @@ def helm_handler(event, context):
             chart_dir = get_chart_from_oci(tmpdir.name, repository, version)
             chart = chart_dir
 
-        helm('upgrade', release, chart, repository, values_file, namespace, version, wait, timeout, create_namespace, atomic=atomic)
+        helm('upgrade', release, chart, repository, values_file, namespace, version, wait, timeout, create_namespace, atomic=atomic, take_ownership=take_ownership)
     elif request_type == "Delete":
         try:
             helm('uninstall', release, namespace=namespace, wait=wait, timeout=timeout)
@@ -156,7 +157,7 @@ def get_chart_from_oci(tmpdir, repository = None, version = None):
     raise Exception(f'Operation failed after {maxAttempts} attempts: {output}')
 
 
-def helm(verb, release, chart = None, repo = None, file = None, namespace = None, version = None, wait = False, timeout = None, create_namespace = None, skip_crds = False, atomic = False):
+def helm(verb, release, chart = None, repo = None, file = None, namespace = None, version = None, wait = False, timeout = None, create_namespace = None, skip_crds = False, atomic = False, take_ownership = False):
     import subprocess
 
     cmnd = ['helm', verb, release]
@@ -181,7 +182,9 @@ def helm(verb, release, chart = None, repo = None, file = None, namespace = None
     if not timeout is None:
         cmnd.extend(['--timeout', timeout])
     if atomic:
-        cmnd.append('--atomic')    
+        cmnd.append('--atomic')
+    if take_ownership:
+        cmnd.append('--take-ownership')
     cmnd.extend(['--kubeconfig', kubeconfig])
 
     maxAttempts = 3
