+++ packages/@aws-cdk/aws-eks-v2-alpha/test/integ.eks-helm-asset.js.snapshot/manifest.json
@@ -18,7 +18,7 @@
         "validateOnSynth": false,
         "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-deploy-role-${AWS::AccountId}-${AWS::Region}",
         "cloudFormationExecutionRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-cfn-exec-role-${AWS::AccountId}-${AWS::Region}",
-        "stackTemplateAssetObjectUrl": "s3://cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}/87840a26e7d59916e46dabec9f741f6b0a0e8a46eabb7c90627b27003ce0d492.json",
+        "stackTemplateAssetObjectUrl": "s3://cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}/b81fa2c67e9ae5a37c31a9f980381d3755ad5d428735356631c2c2799cc85ba5.json",
         "requiresBootstrapStackVersion": 6,
         "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version",
         "additionalDependencies": [
@@ -1054,6 +1054,18 @@
             "data": "Clustercharttestnonecrocichart7A9190B5"
           }
         ],
+        "/aws-cdk-eks-helm-test/Cluster/chart-test-take-ownership-installation/Resource": [
+          {
+            "type": "aws:cdk:analytics:construct",
+            "data": "*"
+          }
+        ],
+        "/aws-cdk-eks-helm-test/Cluster/chart-test-take-ownership-installation/Resource/Default": [
+          {
+            "type": "aws:cdk:logicalId",
+            "data": "ClustercharttesttakeownershipinstallationAE49062F"
+          }
+        ],
         "/aws-cdk-eks-helm-test/LatestNodeRuntimeMap": [
           {
             "type": "aws:cdk:logicalId",
@@ -1166,5 +1178,5 @@
       }
     }
   },
-  "minimumCliVersion": "2.1015.0"
+  "minimumCliVersion": "2.1016.1"
 }
